pyproject.toml
[tool.poetry]
name = "2048neat"
version = "0.1.0"
description = ""
authors = ["Dmytro Maior <mayor4090Gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
neat-python = "^0.92"
torch = "^2.1.0"
torchvision = "^0.16.0"
ipython = "^8.16.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
